Added a cast to remove a compiler warning.
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@4297 23e1bc29-d954-0410-9766-47de34158ba0

+++ PlanDatabase/base/Token.cc
@@ -635,7 +635,7 @@ namespace EUROPA{
     // If the constraint is an inactive subgoal of the master, and its master is committed, and it is strictly in the past
     // then it can be terminated. This is the case of deferred (ignored) subgoals.
     if(isInactive() && getMaster().isId() && getMaster()->isCommitted()){
-      int latestStart = getStart()->lastDomain().getUpperBound();
+      int latestStart = (int) getStart()->lastDomain().getUpperBound();
       if(latestStart < getMaster()->getEnd()->lastDomain().getUpperBound())
 	return true;
     }
